Initiate a major project restructuring
* Removed Readium build tool scripts - Will be migrated over to readium-shared-js
* Removed RequireJS as a bundler and its config files
* Added Rollup.js as a bundler - It's great for producing libraries!
* PEG.js changed to generate in CommonJS module format
* Yarn specifics dropped - It's better to just use one package manager
* Removal of package.cson - The new package.json is simplified so it makes less sense to keep it
* Added ESLint as a linter - Adopting the popular Airbnb style
* Prepare codebase for ES2015+ refactoring
