+++ .editorconfig
@@ -1,0 +1,6 @@
+root = true
+
+[*.js]
+indent_style = space
+indent_size = 2
+max_line_length = 100
