+++ .eslintrc
@@ -1,0 +1,7 @@
+{
+  "env": {
+    "browser": true,
+    "es6": true
+  },
+  "extends": "airbnb-base"
+}
