+++ .gitignore
@@ -1,7 +1,22 @@
 .history
 .DS_Store
-build-output
+dist
 node_modules
 build.txt
 http.log
 api-docs
+
+# JetBrains IDE
+# User-specific stuff:
+.idea/**/workspace.xml
+.idea/**/tasks.xml
+.idea/dictionaries
+
+# Sensitive or high-churn files:
+.idea/**/dataSources/
+.idea/**/dataSources.ids
+.idea/**/dataSources.xml
+.idea/**/dataSources.local.xml
+.idea/**/sqlDataSources.xml
+.idea/**/dynamic.xml
+.idea/**/uiDesigner.xml
