--- dev/index_NO-RequireJS.html
@@ -1,33 +1,0 @@
-<html>
-<head>
-  <!-- script src="http://localhost:9091"> </script -->
-
-<script type="text/javascript" src="../node_modules/jquery/dist/jquery.js"> </script>
-<script type="text/javascript" src="../node_modules/underscore/underscore.js"> </script>
-
-<script type="text/javascript">
-window.EPUBcfi = {};
-</script>
-
-<script type="text/javascript" src="../gen/cfi_parser_gen.js"> </script>
-<script type="text/javascript">
-window.EPUBcfi.Parser = window.EPUBcfiParser;
-window.EPUBcfiParser = undefined;
-</script>
-
-<script type="text/javascript" src="../js/cfi_runtime_errors.js"> </script>
-<script type="text/javascript" src="../js/cfi_instructions.js"> </script>
-<script type="text/javascript" src="../js/cfi_generator.js"> </script>
-<script type="text/javascript" src="../js/cfi_interpreter.js"> </script>
-
-<script type="text/javascript" src="../js/cfi_API.js"> </script>
-
-<script type="text/javascript" src="index.js"> </script>
-
-</head>
-<body>
-  <h1>OPEN CONSOLE TO SEE LOG MESSAGES</h1>
-  
-<div id="viewport"> </div>
-</body>
-</html>
