--- dev/index_RequireJS_multiple-bundles.html
@@ -1,34 +1,0 @@
-<html>
-<head>
-  <!-- script src="http://localhost:9091"> </script -->
-
-<script type="text/javascript" src="../build-output/_multiple-bundles/RequireJS.js"> </script>
-
-<script type="text/javascript" src="../build-output/_multiple-bundles/readium-external-libs.js"> </script>
-
-
-<script type="text/javascript" src="../build-output/_multiple-bundles/readium-cfi-js.js"> </script>
-
-<!-- ============================= -->
-<!-- Demonstrates usage of the "bundles" RequireJS directive, for lazy-loading required dependencies -->
-
-<!--
-
-<script type="text/javascript">
-requirejs.config({
-    baseUrl: '../build-output/_multiple-bundles'
-});
-</script>
-
-<script type="text/javascript" src="../build-output/_multiple-bundles/readium-cfi-js.js.bundles.js"> </script>
--->
-
-
-<script type="text/javascript" src="./index.js"> </script>
-</head>
-<body>
-  <h1>OPEN CONSOLE TO SEE LOG MESSAGES</h1>
-  
-<div id="viewport"> </div>
-</body>
-</html>
