--- dev/index_RequireJS_no-optimize.html
@@ -1,18 +1,0 @@
-<html>
-<head>
-  <!-- script src="http://localhost:9091"> </script -->
-
-<script type="text/javascript" src="../node_modules/requirejs/require.js"> </script>
-
-<script type="text/javascript" src="../build-config/RequireJS_config_browser.js"> </script>
-<script type="text/javascript" src="../build-config/RequireJS_config_common.js"> </script>
-
-<script type="text/javascript" src="index.js"> </script>
-
-</head>
-<body>
-  <h1>OPEN CONSOLE TO SEE LOG MESSAGES</h1>
-  
-<div id="viewport"> </div>
-</body>
-</html>
