--- dev/index_RequireJS_single-bundle.html
@@ -1,14 +1,0 @@
-<html>
-<head>
-  <!-- script src="http://localhost:9091"> </script -->
-
-<script type="text/javascript" src="../build-output/_single-bundle/readium-cfi-js_all.js"> </script>
-
-<script type="text/javascript" src="./index.js"> </script>
-</head>
-<body>
-  <h1>OPEN CONSOLE TO SEE LOG MESSAGES</h1>
-  
-<div id="viewport"> </div>
-</body>
-</html>
