--- dev/index_RequireJS_single-bundle_TESTS.html
@@ -1,40 +1,0 @@
-<html>
-<head>
-  <!-- script src="http://localhost:9091"> </script -->
-
-<script type="text/javascript" src="../build-output/_single-bundle/readium-cfi-js_all.js"> </script>
-
-<script type="text/javascript" src="./index.js"> </script>
-</head>
-<body>
-  <h1>OPEN CONSOLE TO SEE LOG MESSAGES</h1>
-  
-  <!-- The scripts need to be at the end of body, so that some test running frameworks
-   (Angular Scenario, for example) need the body to be loaded so that it can insert its magic
-   into it. If it is before body, then it fails to find the body and crashes and burns in an epic
-   manner. -->
-
-  <script type="text/javascript" src="./index_tests_Karma.js"></script>
-
-  <script type="text/javascript" src="../node_modules/jasmine-core/lib/jasmine-core/jasmine.js"></script>
-<script type="text/javascript" src="../node_modules/karma-jasmine/lib/boot.js"></script>
-<script type="text/javascript" src="../node_modules/karma-jasmine/lib/adapter.js"></script>
-
-<script type="text/javascript" src="../tests/spec/helpers/jasmine-jquery.js"></script>
-<script type="text/javascript" src="index_tests_jasmine-jquery.js"></script>
-
-<script type="text/javascript" src="../tests/spec/models/cfi_generator_spec.js"></script>
-<script type="text/javascript" src="../tests/spec/models/cfi_instructions_spec.js"></script>
-<script type="text/javascript" src="../tests/spec/models/cfi_interpreter_spec.js"></script>
-<script type="text/javascript" src="../tests/spec/models/cfi_parser_spec.js"></script>
-
-<script type="text/javascript" src="../tests/spec/karma-loaded.js"></script>
-
-  <script type="text/javascript">
-  setTimeout(function() {
-    window.__karma__.loaded();
-  }, 1000);
-  </script>
-
-</body>
-</html>
