--- dev/index_tests_jasmine-jquery.js
@@ -1,1 +1,0 @@
-jasmine.getFixtures().fixturesPath = '../tests/spec/fixtures';
