+++ package.json
@@ -1,9 +1,4 @@
 {
-  "/*": "----------------------------------------------------------------------------",
-  "..": "This package.json (JSON) is automatically generated from package.cson (CSON)",
-  ".-": "DO NOT invoke 'npm init' or 'npm install --save' ('--save-dev' / '--save-optional')",
-  "._": "as this would overwrite / update the JSON, not the master CSON!",
-  "*/": "----------------------------------------------------------------------------",
   "name": "readium-cfi-js",
   "version": "0.31.0-alpha",
   "engines": {
@@ -11,6 +6,8 @@
     "npm": ">=2",
     "yarn": ">=0.23"
   },
+  "main": "dist/readium-cfi.umd.js",
+  "module": "dist/readium-cfi.esm.js",
   "description": "Readium - CFI javascript library",
   "keywords": [
     "readium",
@@ -46,57 +43,39 @@
     "README.md"
   ],
   "dependencies": {
-    "concurrently": "^3.5.0",
-    "npm-run-all": "^4.1.2",
-    "opener": "^1.4.3",
-    "http-server": "github:danielweck/http-server#master"
+    "jquery": "2.x.x",
+    "lodash-es": "^4.17.4"
   },
   "devDependencies": {
-    "requirejs": "^2.3.5",
-    "almond": "^0.3.3",
-    "jquery": "2.x.x",
-    "underscore": "^1.8.3",
-    "pegjs": "^0.10.0",
+    "babel-core": "^6.26.0",
+    "babel-plugin-external-helpers": "^6.22.0",
+    "babel-preset-env": "^1.6.1",
+    "eslint": "^4.15.0",
+    "eslint-config-airbnb-base": "^12.1.0",
+    "eslint-plugin-import": "^2.8.0",
+    "jasmine-core": "^2.8.0",
     "karma": "^1.7.1",
     "karma-chrome-launcher": "^2.2.0",
     "karma-firefox-launcher": "^1.0.1",
-    "jasmine-core": "^2.8.0",
     "karma-jasmine": "^1.1.0",
     "karma-sauce-launcher": "^1.2.0",
-    "cson": "^4.1.0",
+    "pegjs": "^0.10.0",
     "rimraf": "^2.6.2",
-    "cpy-cli": "^1.0.1",
-    "ncp": "^2.0.0",
-    "mkdirp": "^0.5.1",
-    "watch": "^1.0.2",
-    "jshint": "^2.9.5",
-    "glob": "^7.1.2",
+    "rollup": "^0.54.0",
+    "rollup-plugin-babel": "^3.0.3",
+    "rollup-plugin-commonjs": "^8.2.6",
+    "rollup-plugin-node-resolve": "^3.0.2",
     "yuidocjs": "^0.10.2"
   },
   "scripts": {
-    "cson2json": "node ./node_modules/cson/bin/cson2json package.cson > package.json && node readium-build-tools/optimizePackageJsonScripts.js",
-    "prepare:all": "npm run prepare:local",
-    "prepare:local": "(npm outdated || echo outdated) && npm install && npm run prepare:local:common",
-    "prepare:yarn:all": "yarn run prepare:yarn:local",
-    "prepare:yarn:local": "(yarn outdated || echo outdated) && yarn install && yarn run prepare:local:common",
-    "prepare:local:common": "node readium-build-tools/patchRequireJS.js && node readium-build-tools/gitHubForksUpdater.js",
-    "clean": "(yarn --version || echo noyarn) && node --version && npm --version && node ./node_modules/rimraf/bin.js build-output/* && node ./node_modules/rimraf/bin.js api-docs/* && node ./node_modules/rimraf/bin.js gen/*",
+    "clean": "rimraf dist/* && rimraf api-docs/* && rimraf gen/*",
     "apidocs": "node ./node_modules/yuidocjs/lib/cli.js -N -C -c yuidocs.json",
-    "prebuild": "npm run clean",
-    "build": "npm run build:scripts",
-    "lint": "node ./readium-build-tools/jshint_glob.js \"js/**.js\"",
-    "prebuild:scripts": "npm run build:scripts:peg",
-    "build:scripts": "npm run build:scripts:multiple && npm run build:scripts:single",
-    "build:scripts:peg": "node ./node_modules/pegjs/bin/pegjs --format globals --export-var window.EPUBcfiParser -o gen/cfi_parser_gen.js cfi_parser.pegjs",
-    "build:scripts:multiple": "node ./node_modules/requirejs/bin/r.js -o ./readium-build-tools/RequireJS_config.js --rjs_bundle=multiple",
-    "postbuild:scripts:multiple": "node ./node_modules/rimraf/bin.js build-output/_multiple-bundles/RequireJS.js.src.js",
-    "build:scripts:single": "node ./node_modules/requirejs/bin/r.js -o ./readium-build-tools/RequireJS_config.js --rjs_bundle=single",
-    "http:watch": "echo > open_webbrowser.js && run-p \"dowatch\" \"http\"",
-    "http": "node readium-build-tools/http-server.js -LOGFILE http.log -OPEN /dev/ -c-1 -a 127.0.0.1 -p 9090 --cors .",
-    "test": "node ./node_modules/karma/bin/karma start tests/karma.conf.js --singleRun --no-auto-watch",
-    "test:travis": "node ./node_modules/karma/bin/karma start tests/karma.travis.conf.js --singleRun --no-auto-watch",
+    "prebuild": "npm run clean && pegjs -o gen/parser.js src/parser.pegjs",
+    "build": "rollup -c",
+    "lint": "eslint src/**/*.js",
+    "test": "karma start tests/karma.conf.js --singleRun --no-auto-watch",
+    "test:travis": "karma start tests/karma.travis.conf.js --singleRun --no-auto-watch",
     "test:travis:firefox": "npm run test:travis --readium-cfi-js:MODE=firefox",
-    "test:travis:chrome": "npm run test:travis --readium-cfi-js:MODE=chrome",
-    "dowatch": "node ./node_modules/watch/cli.js \"npm run build && npm run test && node readium-build-tools/webbrowser.js && node ./node_modules/rimraf/bin.js open_webbrowser.js\" dev/ js/ tests/"
+    "test:travis:chrome": "npm run test:travis --readium-cfi-js:MODE=chrome"
   }
 }
