+++ src/generator.js
