+++ src/interpreter.js
